+++ rust/stackable-cockpit/src/utils/chartsource.rs
@@ -4,6 +4,9 @@ use serde::Deserialize;
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct ChartSourceMetadata {
+    // TODO (@NickLarsenNZ): Add a field for the repo name.
+    // This would remove the need for the nested HashMaps
+    // See: around rust/stackablectl/src/cmds/operator.rs:519
     pub entries: HashMap<String, Vec<ChartSourceEntry>>,
 }